bugfix and split into module / standalone
